requirements.txt
# Core Dependencies
PyQt6==6.6.1
PyQt6-Qt6==6.6.2
PyQt6-sip==13.6.0

# Database
mysql-connector-python==8.2.0
SQLAlchemy==2.0.23

# AI/LLM
groq==0.4.2

# Encryption & Security
cryptography==41.0.7
PyJWT==2.8.1
python-dotenv==1.0.0

# API & Web
requests==2.31.0
httpx==0.25.2

# Database & Storage
python-dateutil==2.8.2

# Email (optional, for notifications)
sendgrid==6.11.0

# Configuration Management
watchdog==3.0.0
pyyaml==6.0.1
tomli==2.0.1

# System Monitoring
psutil==5.9.6

# Development & Testing (optional)
pytest==7.4.3
pytest-cov==4.1.0
bandit==1.7.5
safety==2.3.5
black==23.12.1
flake8==6.1.0
pylint==3.0.3

# Utilities
colorama==0.4.6
tqdm==4.66.1
